Split some of the autocomplete tests out into their own new file (the module related tests)
